+++ connector-packager/connector_packager/connector_properties.py
@@ -7,3 +7,4 @@ class ConnectorProperties:
         self.backwards_compatibility_mode = False
         self.is_jdbc = False
         self.connection_metadata_database = True
+        self.oauth_config_ids = []