+++ connector-packager/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.2] - 4/9/2026
+- Add check for non-lowercase default OAuth config IDs
+- Add check for duplicate OAuth config IDs
+- Add check for missing default OAuth config ID if non-default IDs are defined
+
 ## [2.1.1] - 8/29/2025
 - Add checks for secure field violations